requirements.txt
tensorflow==2.5.0
streamlit==0.85.1
pandas==1.3.1
tensorflow_datasets==4.4.0
matplotlib==3.4.2
Pillow==8.3.1
